requirements.txt
# requirements.txt
Flask==2.2.3
Flask-PyMongo==2.3.0
Flask-SocketIO==5.3.3
gunicorn==20.1.0
pymongo==4.3.3
requests==2.28.2
python-dotenv==1.0.0
eventlet==0.33.3
dnspython==2.3.0
python-engineio==4.4.1
python-socketio==5.8.0
